+++ .gitignore
@@ -7,3 +7,5 @@ raw_data/
 **/__pycache__/
 .env
 .python-version
+*.joblib
+.joblib
